stargazer_style_list {stargazer}	R Documentation

stargazer: list of supported styles

Description

The following character strings can be used in the style argument of the stargazer command. Most styles are designed to resemble an existing academic journal, as listed below.

"all" 	 publish every statistic available, incl. t-statistics and p-values

"all2" 	 same as "all", but omitting t-statistics and p-values

"default" 	 default: publish regression coefficients with standard errors, and the most commonly reported statistics

"commadefault" 	 like "default", but uses a decimal comma and a single space to separate thousands

"aer" 	 American Economic Review

"ajps" 	 American Journal of Political Science

"ajs" 	 American Journal of Sociology

"asq" 	 Administrative Science Quarterly

"asr" 	 American Sociological Review

"apsr" 	 American Political Science Review

"demography" 	 Demography

"io" 	 International Organization

"jpam" 	 Journal of Policy Analysis and Management

"qje" 	 Quarterly Journal of Economics

[Package stargazer version 5.2 Index]