stargazer_news {stargazer}	R Documentation

What's new?

Description

This page summarizes the models, features, and bug fixes that were introduced in each release of stargazer. Please note that I am always looking for comments and suggestions. Do not hesitate to contact me at hlavac [at] fas.harvard.edu.

What's new?

v. 5.2:

- New models:
– arima (stats)
– censReg (censReg)
– garchFit (fGarch)
– glmrob (robustbase)
– lme (nlme)
– nlme (nlme)
– mnlogit (mnlogit)
– pgmm (plm)

- New features:
– decimal mark (argument decimal.mark) now gets automatically updated in notes (argument notes)

- Fixed bugs:
– felm (lfe) working again
– fixed formatting issues that led to an occasional inability to produce ASCII text output
– much faster output of multi-model tables with many omitted variables
– negative numbers are now properly displayed when initial.zero is set to FALSE
– fixed minor bug that occasionally led to a misalignment of additional lines (argument add.lines)
– lm() and glm() models now report AIC and BIC
– underscores are now properly formatted in column and dependent variable labels
– fixed an issue with the mlogit model that led LR tests to crash the output when no intercept was present

v. 5.1:

- New features:
– the CRAN description now lists packages whose usefulness is enhanced by stargazer

- Fixed bugs
– labels that indicate whether variables have been omitted from the model are now correctly assigned (argument omit.labels)
– argument summary.stat now functions properly
– out.header and header can both be set to false at the same time
– minor formatting changes

v. 5.0:

- New models:
– felm (lfe)

- New features:
– ability to produce HTML/CSS output (argument type)
– vector and matrix objects are now acceptable (argument ...)
– additional lines can be included in the tables (argument add.lines)
– table layout is now fully customizable (arguments table.layout and omit.table.layout)
– choice over which summary statistics should be reported (arguments summary.stat and omit.summary.stat)
– users can choose whether and the order in which to report coefficients, standard errors, confidence intervals, test statistics and p-values (argument report)
– object names can be reported above each column (argument object.names)
– option to toggle the reporting of dependent variables and model names across several columns (argument multicolumn)
– reporting of model statistics' degrees of freedom can now be suppressed (argument df)
– users can now include more than three statistical significance cutoffs (argument star.cutoffs)
– regression, summary statistic and data frame table can now be flipped (argument flip)
– colnames and rownames can be included in, or excluded from, data frame content tables (arguments colnames and rownames)
– argument summary can now have a different value for each table within a single stargazer() call
– if arguments coef, se, t, p, ci.custom and add.lines only contain a single vector, they will be accepted as though they were a list containing the same vector

- Fixed bugs:
– fixed bug that caused the number of observations to be misreported in complicated models
– fixed bug that led to the omission of coefficients from output when multiple models had a single regressor
– notes now working well for summary statistics and data frame tables
– font size can now be change even if floating environment is not used
– fixed compatability issues with latest version of Zelig
– argument apply.ci now works properly
– minor formatting changes

v. 4.5.3:

- New features:
– model objects can now be passed to stargazer in a list (argument ...)

- Fixed bugs:
– fixed problem with printing negative numbers in data frame and summary statistics tables
– corrected formatting of column widths in ASCII text output
– minor bug with lme4 covariate names corrected

v. 4.5.2:

- Fixed bugs:
– much faster output
– now compatible with the updated lme4 package
– corrected a bug that prevented users from including multiple models with a single covariate + intercept
– underscores now print correctly in ASCII text output

v. 4.5.1:

- Fixed bugs:
– fixed a bug involving the calculation of t-statistics from user-given coefficients and standard errors

v. 4.5:

- New models:
– binaryChoice (sampleSelection)
– brglm (brglm)
– coeftest (lmtest)
– heckit (sampleSelection)
– maBina (erer)
– mclogit (mclogit)
– mlogit (mlogit)
– selection (sampleSelection)

- New features:
– explanatory variables can be ordered in customizable ways (argument order)
– custom functions can now be applied to coefficients (argument apply.coef), standard errors (argument apply.se), test statistics (argument apply.t), p-values (argument apply.p), and confidence intervals (argument apply.ci)
– keep explanatory variables and statistics based on regular expressions (argument keep and keep.stat)
– users can customize confidence intervals (argument custom.ci)
– ability to have both standard errors and confidence intervals in the same table (argument ci is now a logical vector)
– different confidence levels for each individual column (argument ci.level is now a vector)
– ability to omit and keep variables based on their position in the table

- Fixed bugs:
– citation and change log updates
– minor formatting changes

v. 4.0:

- New models:
– aftreg (eha)
– bj (rms)
– coxreg (eha)
– cph (rms)
– dynlm (dynlm)
– errorsarlm (spdep)
– Glm (rms)
– Gls (rms)
– gmm (gmm)
– hetglm (glmx)
– lrm (rms)
– mlreg (eha)
– lagsarlm (spdep)
– ols (rms)
– phreg (eha)
– psm (rms)
– rem.dyad (relevent)
– rq (quantreg)
– Rq (rms)
– weibreg (eha)

- New features:
– the package can produce ASCII text output, in addition to LaTeX code (argument type)
– output directly to .tex or .txt files (argument out)
– column labels (arguments column.labels and column.separate)
– confidence intervals (arguments ci, ci.level and ci.separator)
– coefficients and standard errors/confidence intervals can now be reported in the same row (argument single.row)
– users can choose to omit all empty lines in a table (argument no.space)
– notes can now be appended to, rather than always replace, the default notes for a given style (argument notes.append)
– ability to customize the dependent variable caption (argument dep.var.caption)
– font size can now be changed (argument font.size)
– comments header (with package and author name, version, date and time) can now be suppressed (argument header)
– ability to change or disable the floating environment (arguments float and float.env)
– table placement settings (argument table.placement)
– customization of column spacing (argument column.sep.width)
– perl-compatible regular expressions (argument perl)

- Fixed bugs:
– all columns are now displayed for multinom objects
– better positioning of tables in the LaTeX document
– minor formatting changes

v. 3.0.1:

- Fixed bugs:
– corrected an issue that led to problems when printing data frames with multiple decimal places
– some reporting and formatting changes to summary statistics and data frame tables

v. 3.0:

- New models:
– clm (ordinal)
– clogit (survival)
– ergm (ergm)
– glmer (lme4)
– gls (nlme)
– lmer (lme4)
– lmrob (robustbase)
– nlmer (lme4)
– pmg (plm)
– rlm (MASS)

- New features:
– users can customize coefficients (argument coef), standard errors (se), test statistics (t), and p-values (p)
– automatic calculation of z-scores and p-values when the user supplies custom standard errors (can be toggled using arguments t.auto and p.auto)
– ability to set \label{} markers in TeX for each table (using argument label)
– summary statistics table can now report logical (i.e., dummy, indicator) variables as if they had values of 0 (corresponding to FALSE) and 1 (TRUE). See argument summary.logical.
– user can choose between rounding to a set number of decimal places (argument digits) and reporting all available decimal places (by setting digits equal to NA)
– can omit all test statistics

- Fixed bugs:
– package runs much faster
– corrected p-values for polr() and zelig oprobit, ologit models
– coxph now, by default, reports robust standard errors, for consistency with its summary() output
– multinom models now report the first set of coefficients (when multiple sets are present), instead of producing an error
– singularities in model regression no longer disrupt the output
– model formulas can now be symbols
– no more warnings when creating summary statistics tables
– minor formatting changes

v. 2.0.1:

- Fixed bugs:
– stargazer now reports the correct number of observations for plm() models

v. 2.0:

- New models:
– betareg (betareg)
– hurdle (pscl)
– ivreg (AER)
– multinom (nnet)
– plm (plm)
– tobit (AER)
– zeroinfl (pscl)

- New features:
– direct output of data frames into LaTeX (summary = FALSE)
– ability to omit the reporting of selected statistics (using argument omit.stat) – alignment of columns at the decimal mark (align = TRUE)
– automatic coordination of star cutoffs and values in regression table notes

- Fixed bugs:
– argument digits now works properly
– stargazer updated to reflect, and work properly with, recent changes to Zelig
– variable names can now contain dollar signs and underscores
– some minor formatting fixes

[Package stargazer version 5.2 Index]